requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
attrs==22.1.0
autopep8==1.6.0
black==22.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
coverage==6.4.2
db-dtypes==1.0.0
flake8==4.0.1
google-api-core==2.7.1
google-auth==2.6.3
google-cloud-bigquery-storage==2.13.1
google-cloud-bigquery==3.0.1
google-cloud-core==2.3.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio-status==1.44.0
grpcio==1.44.0
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3; python_version >= '3.6'
pandas==1.4.2
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.20.3
protobuf==3.20.0
py==1.11.0
pyarrow==7.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
pygments==2.13.0
pyparsing==3.0.8
pytest-cov==3.0.0
pytest-mock==3.8.2
pytest-responses==0.5.0
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pyyaml==6.0
requests-mock==1.9.3
requests==2.27.1
responses==0.21.0
rich==12.5.1
rsa==4.8
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.9
